requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
flask-socketio
flask-cors
firebase-admin
flask-login
pandas
matplotlib
passlib
eventlet
gunicorn
python-dotenv